+++ CHANGELOG.md
@@ -1,3 +1,9 @@
-# 1.0.4
+# 1.0.5
 
-对外版本发布
+## BugFix
+- 修复build和deploy时不清空目录的BUG
+- 修复build只有一个页面时候，watch进程不退出的问题
+- 修复自动更新提示Changelog链接错误问题
+
+## Feature
+- 增加了快捷创建tsconfig的命令`muse create tsconfig`